+++ site/src/routes/sheet-demo/SheetDemoRoute.tsx
@@ -1,0 +1,5 @@
+import { SheetDemo } from "./SheetDemo";
+
+export function SheetDemoRoute() {
+  return <SheetDemo />;
+}